Moved Server Side logic to lib folder and lint clean
