+++ lib/initial-console.ts
