+++ pages/index.tsx
@@ -1,4 +1,4 @@
-import { initialConsole } from './api/initial-console';
+import { initialConsole } from '../lib/initial-console';
 
 export const getStaticProps = () => {
   initialConsole('Home')
