+++ src/utils/InitGetSession.ts
@@ -13,7 +13,7 @@ const InitGetSession = () => {
     if(data?.session) {
       dispatch(update_userData(data.session))
     }
-  }, [dispatch])
+  }, [dispatch, supabaseClient.auth])
 
   useEffect(() => {
     updateInitialUserData()
